Moved plugin state test to client crate
