+++ packages/client/tests/it/wrap_features/mod.rs
@@ -3,3 +3,5 @@ pub mod env_with_subinvoke;
 pub mod interface_implementation;
 pub mod plugin_invocation;
 pub mod subinvoke;
+
+pub mod plugin_state;
