+++ packages/plugin/Cargo.toml
@@ -22,9 +22,6 @@ serde.workspace = true
 bigdecimal.workspace = true
 num-bigint.workspace = true
 
-[dev-dependencies]
-polywrap_client_builder.workspace = true
-
 [features]
 default = ["implementor"]
